Buttons/Display Working for Single Calculations
When a number or operator button is pressed, the display updates and the
x, y, and operator values are set accordingly. Can run the Operate
function fine once. Still need to work on being able to do operate
multiple times in a row along with clearing all the variable on the clear
button press

[Ticket: X]